requirements.txt
pymysql
python-dotenv
pandas
sqlalchemy